gcf-test/Cloud_run_functions/HTTPTriggers/Bigquery_webpage/requirements.txt
flask==2.3.3
functions-framework==3.5.0
google-cloud-bigquery==3.12.0
google-cloud-secret-manager==2.8.0
google-cloud-storage==1.45.0
google-cloud-logging==2.8.0